requirement.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles

mistralai

langchain
langchain-community
langchain-text-splitters
langchain-mistralai

faiss-cpu
pypdf

python-dotenv
pydantic
tqdm

pandas
numpy